+++ packages/decap-cms-backend-git-gateway/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.6.0](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-git-gateway@3.5.1...decap-cms-backend-git-gateway@3.6.0) (2026-04-16)
+
+### Features
+
+- add config option to enable Signed-off-by in commit messages ([#7766](https://github.com/decaporg/decap-cms/issues/7766)) ([4b9ea40](https://github.com/decaporg/decap-cms/commit/4b9ea4040f3b858793052cee287caa3cce2f3af3))
+
+## [3.5.1](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-git-gateway@3.5.0...decap-cms-backend-git-gateway@3.5.1) (2026-02-23)
+
+### Bug Fixes
+
+- `import/no-named-as-default` warning ([#7721](https://github.com/decaporg/decap-cms/issues/7721)) ([f7ef707](https://github.com/decaporg/decap-cms/commit/f7ef70717aeaa449ef32624c493f340e14501b68))
+
 # [3.5.0](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-git-gateway@3.4.1...decap-cms-backend-git-gateway@3.5.0) (2025-12-18)
 
 **Note:** Version bump only for package decap-cms-backend-git-gateway