+++ packages/decap-cms-widget-markdown/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.8.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-markdown@3.7.0...decap-cms-widget-markdown@3.8.0) (2026-04-16)
+
+### Bug Fixes
+
+- **markdown:** extra whitespace before and after pasted content [#7364](https://github.com/decaporg/decap-cms/issues/7364) ([#7440](https://github.com/decaporg/decap-cms/issues/7440)) ([fc16ff3](https://github.com/decaporg/decap-cms/commit/fc16ff3d303ddf0840c3dacea7f34449bb69f2f3))
+
+### Features
+
+- remove HTML comments from pasted content ([#7775](https://github.com/decaporg/decap-cms/issues/7775)) ([b7cf396](https://github.com/decaporg/decap-cms/commit/b7cf3968b157f13da9ca6c9448ab8e8bc51a4afa))
+
+# [3.7.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-markdown@3.6.1...decap-cms-widget-markdown@3.7.0) (2026-03-24)
+
+### Bug Fixes
+
+- **markdown:** list editor components alphabetically in toolbar ([#7759](https://github.com/decaporg/decap-cms/issues/7759)) ([81cc055](https://github.com/decaporg/decap-cms/commit/81cc05578a88d338361581533845eca4c4eed3a6))
+
+## [3.6.1](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-markdown@3.6.0...decap-cms-widget-markdown@3.6.1) (2026-02-23)
+
+### Bug Fixes
+
+- **a11y:** aria-labels ([#7720](https://github.com/decaporg/decap-cms/issues/7720)) ([cc92370](https://github.com/decaporg/decap-cms/commit/cc92370595a85590ffede50a99259d52ba3eb76b))
+- **markdown:** remove empty links when text is deleted ([#7702](https://github.com/decaporg/decap-cms/issues/7702)) ([f1f1d13](https://github.com/decaporg/decap-cms/commit/f1f1d13e9c3f4b6909371c6fc2e74fe0204d7029)), closes [#7640](https://github.com/decaporg/decap-cms/issues/7640)
+
 # [3.6.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-markdown@3.5.0...decap-cms-widget-markdown@3.6.0) (2025-11-04)
 
 ### Bug Fixes